+++ rdiffweb/plugins/notification/__init__.py
@@ -38,7 +38,8 @@ from rdiffweb import librdiff
 from rdiffweb.core import RdiffError
 from rdiffweb.i18n import ugettext as _
 from rdiffweb.rdw_helpers import rdwTime
-from rdiffweb.rdw_plugin import IPreferencesPanelProvider, JobPlugin
+from rdiffweb.rdw_plugin import IPreferencesPanelProvider, JobPlugin, \
+    IUserChangeListener
 
 
 _logger = logging.getLogger(__name__)
@@ -113,7 +114,7 @@ def _utf8(self, val):
     return val.encode('utf-8')
 
 
-class NotificationPlugin(IPreferencesPanelProvider, JobPlugin):
+class NotificationPlugin(IPreferencesPanelProvider, JobPlugin, IUserChangeListener):
     """
     Send email notification when a repository get too old (without a backup).
     """
@@ -153,15 +154,38 @@ class NotificationPlugin(IPreferencesPanelProvider, JobPlugin):
         """Called by the plugin manager to setup the plugin."""
         super(NotificationPlugin, self).activate()
 
+        # If notification are disabled, delete the handler.
+        notify = self.app.cfg.get_config_bool('EmailSendChangedNotification', False)
+        if not notify:
+            del self.user_attr_changed
+
     @property
     def job_execution_time(self):
+        """
+        Implementation of JobPlugin interface.
+        """
         return self.app.cfg.get_config('EmailNotificationTime', '23:00')
 
     def job_run(self):
         """
+        Implementation of JobPLugin interface.
         Go trough all the repositories and users to send mail in batches.
         """
         self.send_notifications()
+
+    def user_attr_changed(self, username, attrs={}):
+        """
+        Implementation of IUserChangeListener interface.
+        """
+        # Leave if the mail was not changed.
+        if 'email' not in attrs:
+            return
+
+        # get User object (to get email)
+        userobj = self.app.userdb.get_user(username)
+        assert userobj
+        # If the email attributes was changed, send a mail notification.
+        self.send_mail(userobj, _("Email address changed"), "email_changed.html")
 
     def send_notifications(self):
         """